refactor(types): Reorganize type files and update import paths
The type definition files have been restructured for better organization. The generic `types.ts` has been moved to a more specific `worldbook.ts`, and character-related types are now grouped under `types/character/`.

All corresponding import paths have been updated throughout the application to use the `@` path alias instead of relative paths. This makes the imports cleaner, more consistent, and more resilient to future file movements.
