+++ src/composables/worldbook/useWorldBookDragDrop.ts
@@ -1,5 +1,5 @@
 import type { Ref } from 'vue';
-import type { WorldBookCollection, WorldBookEntry } from '../../types/types';
+import type { WorldBookCollection, WorldBookEntry } from '@/types/worldbook';
 import type { AllowDropType, NodeDropType } from 'element-plus/es/components/tree/src/tree.type';
 import { ElMessage } from 'element-plus';
